requirements.txt
amqp==5.3.1
anyio==4.8.0
APScheduler==3.6.3
arrow==1.3.0
asgiref==3.8.1
billiard==4.2.1
blessed==1.20.0
cachetools==4.2.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.1
defusedxml==0.7.1
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.4.1
django-filter==25.1
django-jazzmin==3.0.1
django-picklefield==3.2
django-q==1.3.9
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huey==2.5.2
idna==3.10
kombu==5.4.2
oauthlib==3.2.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-telegram-bot==13.15
python3-openid==3.2.0
pytz==2025.1
redis==3.5.3
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
telebot==0.0.5
tornado==6.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
vine==5.1.0
vulture==2.14
wcwidth==0.2.13
whitenoise==6.9.0
gunicorn==23.0.0
django-cors-headers==4.7.0
